Откат до ревизии cb6607859af82a8bbee9e98d46f99a883289814c

+++ internal/server/storage/dbStorage.go
@@ -9,7 +9,7 @@ func createTables(db *sql.DB) error {
 	sqlStatement := `
 CREATE TABLE IF NOT EXISTS counter (
     name VARCHAR(64),
-    value integer,
+    value NUMERIC,
     PRIMARY KEY (name)
 );
 CREATE TABLE IF NOT EXISTS gauge (
